+++ include/neoSYCL/sycl/detail/device.hpp
@@ -47,7 +47,7 @@ struct device_impl_host : public device_impl {
     switch (param) {
     case info::device::device_type:
     default:
-      PRINT_ERR("device::get_info(%d) not implemented", param);
+      PRINT_ERR("device::get_info(%d) not implemented", (int)param);
       throw unimplemented();
     }
   }
